requirements.txt
pytorch
torchvision
numpy
tqdm
mlflow
imageio